requirements.txt
beautifulsoup4
bertopic
hdbscan
joblib
nltk
openpyxl
pandas
praw
python-dotenv
scikit-learn
sentence-transformers
spacy
tqdm
umap-learn
streamlit
plotly
matplotlib
wordcloud
